statistics: fix conntrack regression caused by collectd 5.5.0
Collectd 5.5.0 introduced new data to conntrack plugin:
In addition to the number of tracked connections there is also
the static max conntrack value and the calculated use percentage.

Luci's conntrack plugin intrepretes "conntrack-max" as a new data instance
and includes it in the graph in addition to the real "conntrack" number.

Eliminate "max" from graph by specifying empty "" instance as data source.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>