statistics: cpu graph - add label definitions, add softirq and interrupt stats
CPU plugin in the Luci statistics was missing the label definitions,
so the field labels are like "cpu_system" instead of "System".
Add proper label definitions to CPU (like the other plugins already have).

The statistics graph was also missing softirq and interrupt stats, although colors
for them were defined. Softirq consumes massive amount of CPU especially with
any qos in use, so it is important for the user to see also that data. Add both
softirq and interrupt stats to the graph.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>