luci-base: set default mediaurlbase to bootstrap (current default theme)
The mediaurlbase option in the default /etc/config/luci still points
to the old openwrt.org theme that is not installed by default.
The discrepancy was noted in the commit message for 55ab4e4ce22

After 55ab4e4ce22 the installed theme's uci-defaults script will correct
the setting at first boot, but we should not have a deprecated theme as
the default value. Set the default value to the default theme 'bootstrap'.

Related old discussion at #302

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>