luci-app-pbx(-voicemail): remove from repo
Remove the luci-app-pbx(-voicemail) packages, as they have been
marked BROKEN since August 2016.

The pbx packages depend on asterisk 1.8 that is EOL upstream
and has been moved from the telephony feed to the abandoned feed.

If LuCI pbx packages are still needed, they should be refreshed
to depend on a current asterisk version from the telephony feed,
and to match the features of that asterisk version.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>