luci-app-snmpd: Allow turning off agentx
AgentX support doesn't actually require TCP, it also works over unix
domain sockets, and UDS is the only method that's compiled in by
default.  Remove that misleading text, and make the section
add/removable so that you can remove it to disable agentx support.
Behaviour with multiple sections is undefined. (don't do that!)

This matches the current behaviour of the snmpd init script, which will
enable agentx with the compile time default settings if the agentx
socket config is blank/missing.

Signed-off-by: Karl Palsson <karlp@etactica.com>