luci-app-nut: Update settings to match packages master
A number of settings in openwrt-packages master are missing; add them.
The 'extra' adhoc options (override, default, other, and otherflag) we
leave to UCI only because they require dynamic creation of config
sections.
There were some options that didn't belong in the section they were
in; remove them.
Include the option for the USB serial port permissions for serial USB
devices over USB serial port (on the OpenWrt device).
And finally, fix the title of the UPS Server global settings.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>