luci-app-battstatus: initial release
Introduce a new package luci-app-battstatus which queries battery charge
information using i2c commands and displays a charge indicator in the
upper right notification area.

So far this package has been tested to work with the HooToo HT-TM05
travel router but might be extended to cover further models in the
future.

Signed-off-by: Russell Morris <rmorris@rkmorris.us>
[rebase on top of master, rewrite commit message, adjust copyright,
 convert some ubus fields to boolean values]
Signed-off-by: Jo-Philipp Wich <jo@mein.io>