luci-app-sqm: Fix run path
luci-app-sqm was using the wrong path to look for the available_qdiscs
file. This happened to work because by default /var is symlinked to /tmp on
OpenWrt. However, it breaks badly when this is not the case, and the
sqm.conf file for OpenWrt uses /var/run anyway, so update the luci code to
do this as well.

Fixes #3860.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>